\name{book}
\alias{book}
\docType{data}
\title{Books about US politics}
\description{
A network of books about US politics published 
around the time of the 2004 presidential election 
and sold by the online bookseller Amazon.com. Edges between books represent frequent copurchasing of books by the same buyers.
}
\usage{data(book)}
\format{
The data are in the format of igraph object.
IGRAPH U--- 105 441 -- 
+ attr: layout (g/n), randomout (g/n), id (v/n), label (v/c), value (v/c)
}
\source{
http://www-personal.umich.edu/~mejn/netdata/
http://www.orgnet.com/
}
\examples{
data(book)
}
\keyword{datasets}
