\name{karate}
\alias{karate}
\docType{data}
\title{
Karate Club Data
}
\description{
Social network of friendships between 34 members of a karate club at a US university in the 1970s
}
\usage{data(karate)}
\format{
The data are in the format of igraph object.
IGRAPH U--- 34 78 -- 
+ attr: layout (g/n)
}
\source{
http://www-personal.umich.edu/~mejn/netdata/
}
\references{
Zachary, W. W. (1977). An information flow model for conflict and fission in small groups. 
\emph{Journal of anthropological research}, 452--473.
}
\examples{
data(karate)
}
\keyword{ datasets }
