\name{SNscan-package}
\alias{SNscan-package}
\alias{SNscan}
\docType{package}
\title{Scan Statistics in Social Networks}
\description{
This package is constructed for testing the clustering patterns of structure and attribute 
among a social network through the scan statistics. 
Most contents are related to the references mentioned below.
Some data sets are presented to be examples in this package as well.
}
\details{
\tabular{ll}{
Package: \tab SNscan\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-1-9\cr
}
}
\references{
Wang, T. C., & Phoa, F. K. H. (2016). 
A scanning method for detecting clustering pattern of both attribute and structure in social networks. 
Physica A: Statistical Mechanics and its Applications, 445, 295-309.\cr
Wang, T. C., Phoa, F. K. H., & Hsu, T. C. (2015). Power-law distributions of attributes in community detection. Social Network Analysis and Mining, 5(1), 1-10.
}
\author{
Tai-Chi Wang [aut,cre,ths] <taichi53719@gmail.com>, 
Tun-Chieh Hsu [aut] <htj801116@gmail.com>, 
Frederick Kin Hing Phoa [ths] <fredphoa@stat.sinica.edu.tw>
}
\keyword{ package }
\seealso{
\code{\link[igraph]{igraph}}
}

