\name{zetatable}
\alias{zetatable}
\docType{data}
\title{
Zeta Function Table 
}
\description{
Each row of zeta function table equals to the derivation of Riemann zeta function divided by zeta function.
}
\usage{data(zetatable)}
\format{
  A data frame with 99999 observations on the following 2 variables;
  The first column is the given value of alpha in zeta function,
  and the second is the value of the derivation of Riemann zeta function divided by zeta function.
}
\examples{
data(zetatable)
str(zetatable)
}
\keyword{datasets}
